adding multidimensional

+++ c/floats-cbmc-regression/float_lib1_true-unreach-call.c
@@ -12,7 +12,7 @@ int main() {
   #else
   // Be careful, result of DBL_MAX+DBL_MAX depens on rounding mode!
   if(!(fpclassify(DBL_MAX+DBL_MAX)==FP_INFINITE)) __VERIFIER_error();
-  if(!(fpclassify(0*(DBL_MAX+DBL_MAX))==FP_NAN)) __VERIFIER_error();
+  if(!(fpclassify(0*(LDBL_MAX+LDBL_MAX))==FP_NAN)) __VERIFIER_error();
   if(!(fpclassify(1.0)==FP_NORMAL)) __VERIFIER_error();
   if(!(fpclassify(DBL_MIN)==FP_NORMAL)) __VERIFIER_error();
   if(!(fpclassify(DBL_MIN/2)==FP_SUBNORMAL)) __VERIFIER_error();
